move stories into docs folder
