+++ .gitignore
@@ -1,3 +1,3 @@
 dist
-docs
 node_modules
+/docs
