+++ packages/ui-modules-canvas/docs/stories.jsx
@@ -1,15 +1,13 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { withKnobs, text } from '@storybook/addon-knobs';
+import { withKnobs, number } from '@storybook/addon-knobs';
 /* eslint-enable */
 
-import './style.css';
+import '../style.css';
 
-storiesOf('Footer/1.0.0', module)
+storiesOf('Canvas/1.0.0', module)
   .addDecorator(withKnobs)
   .addWithStaticMarkup('default', () => (
-    <footer className="footer" role="contentinfo">
-      {text('Content', 'Footer')}
-    </footer>
+    <canvas width={number('width', '800')} height={number('height', '600')} />
   ));
