+++ packages/ui-modules-headline/package.json
@@ -2,7 +2,7 @@
   "name": "@marcobiedermann/ui-modules-headline",
   "version": "1.0.0",
   "description": "",
-  "style": "index.css",
+  "style": "style.css",
   "scripts": {
     "build": "node ../../scripts/build.js",
     "lint:css": "stylelint *.css",
